- Fjernet unødvendig isFetching (mulig denne må kanskje benyttes igjen ved fil parsing) - Rekursjon OK
